tests/pyproject.toml
[tool.poetry]
name = "tests"
version = "1.0.0"
description = "test grpc app"
authors = ["Kirill <teckytgut@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio-tools = "^1.60.1"
grpcio = "^1.60.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
